refactor: move request id and global exception to plugin
