+++ apps-api/backoffice/package.json
@@ -30,7 +30,6 @@
     "@sinclair/typebox": "^0.34.37",
     "dotenv": "^17.0.1",
     "elysia": "^1.3.5",
-    "elysia-ip": "^1.0.10",
     "jsonwebtoken": "^9.0.2",
     "neverthrow": "^8.2.0",
     "pino": "^9.7.0",
